# docs/swagger.yaml
basePath: /
definitions:
...
  dtos.RewardRequest:
    properties:
      description:
        type: string
      points_required:
        type: number
      store_id:
        type: integer
    type: object
  dtos.StoreRequest:
    properties:
      conversion_factor:
        type: number
      name:
        type: string
    type: object
  dtos.TransactionRequest:
    properties:
      amount:
        type: number
      branch_id:
        type: integer
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: API test.
  title: mi api
  version: "1.0"
paths:
...
  /leal-test/rewards:
    get:
      consumes:
      - application/json
      description: Get all rewards
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all rewards
      tags:
      - rewards
    post:
      consumes:
      - application/json
      description: Create a new reward
      parameters:
      - description: Reward data
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/dtos.RewardRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new reward
      tags:
      - rewards
  /leal-test/rewards/{id}:
    delete:
      description: Delete a reward by ID
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete reward by ID
      tags:
      - rewards
    get:
      consumes:
      - application/json
      description: Get reward by ID
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get reward by ID
      tags:
      - rewards
    put:
      consumes:
      - application/json
      description: Update a reward
      parameters:
      - description: Reward ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reward data
        in: body
        name: reward
        required: true
        schema:
          $ref: '#/definitions/dtos.RewardRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update a reward
      tags:
      - rewards
  /leal-test/rewards/store/{store_id}:
    get:
      consumes:
      - application/json
      description: Get rewards by StoreID
      parameters:
      - description: Store ID
        in: path
        name: store_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get rewards by StoreID
      tags:
      - rewards
  /leal-test/stores:
    get:
      consumes:
      - application/json
      description: Get all stores
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all stores
      tags:
      - stores
    post:
      consumes:
      - application/json
      description: Create store
      parameters:
      - description: Store to create
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/dtos.StoreRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create store
      tags:
      - stores
  /leal-test/stores/{id}:
    delete:
      consumes:
      - application/json
      description: Delete store by ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete store by ID
      tags:
      - stores
    get:
      consumes:
      - application/json
      description: Get store by ID
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get store by ID
      tags:
      - stores
    put:
      consumes:
      - application/json
      description: Update store
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Store to update
        in: body
        name: store
        required: true
        schema:
          $ref: '#/definitions/dtos.StoreRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update store
      tags:
      - stores
  /leal-test/transactions:
    get:
      consumes:
      - application/json
      description: Get all transactions
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get all transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dtos.TransactionRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /leal-test/transactions/{id}:
    get:
      consumes:
      - application/json
      description: Get transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
  /leal-test/transactions/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get transactions by UserID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get transactions by UserID
      tags:
      - transactions
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"